+++ CHANGELOG.md
@@ -7,9 +7,12 @@ This is a manually written changelog, and only tracks front-facing changes since
 - `download`: change editors to run with inherited STDIO
 - `download`: add ability to play the current element in the edit stage
 - `download`: add ability to go back elements in the edit stage
+- `download`: change to use command `yt-dlp` instead of `youtube-dl`
 - ffmpeg: fix error on invalid utf8 sequence
 - change Termination requests to not be based on time anymore
 - some more internal refactoring
+- build requirements section has been added
+- requirements have been updated
 
 ## v0.7.0